backend/requirements.txt
Django==5.1.7
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0
